+++ R/pdf_pages.R
@@ -2,21 +2,13 @@
 #'
 #' `pdf_pages()` calculates an integer vector of subset of pdf pages.
 #'
-#' @inheritParams pdf_pad
+#' @inheritParams pdf_pad_pagesize
 #' @param pages A positive numeric vector of pages to include,
 #'              a negative numeric vector of pages to exclude,
 #'              or "all", "even", or "odd" to include all pages, just the even pages,
 #'              or just the odd pages.
 #' @examples
-#' # Create an 8-page document
-#' f <- tempfile(fileext = ".pdf")
-#' grDevices::pdf(f)
-#' for (page in seq.int(8L)) {
-#'   grid::grid.newpage()
-#'   grid::grid.text(paste("Page", page))
-#' }
-#' invisible(grDevices::dev.off())
-#'
+#' f <- pdf_blank(length = 8L)
 #' pdf_pages(f, pages = 1:4)
 #' pdf_pages(f, pages = -(1:4))
 #' pdf_pages(f, pages = "all")