+++ NEWS.md
@@ -13,7 +13,8 @@ New features
   + `pdf_compress()`, `pdf_rotate_pages()`, and `pdf_subset()` wrap the `{qpdf}` functions of the
     same name but follow the conventions of this package (i.e. input first positional argument, output second positional argument and defaults to temporary file, all other arguments must be named).
   + `pdf_gs()` processes a pdf file with ghostscript.
-  + `pdf_pad()` makes a pdf file larger by padding it (i.e. adding space to the outside margins).
+  + `pdf_pad_length()` adds blank pages to the end of a pdf file (#10).
+  + `pdf_pad_pagesize()` makes a pdf file larger by padding it (i.e. adding space to the outside margins).
     The original images are **not** rescaled (#5).
   + `pdf_rm_crosshairs()` removes crosshairs from a pdf file.
     Currently only supports removing crosshairs from
@@ -22,6 +23,7 @@ New features
 Other utility functions:
 
   + `ls_temp_pdfs()` lists temporary pdfs while `rm_temp_pdfs()` removes temporary pdfs.
+  + `pdf_blank()` creates blank pdf files.
   + `pdf_pages()` calculates an integer vector of subset of pdf pages (#9).
   + `pdf_width()` and `pdf_height()` get pdf page dimensions.